Changes done as per comment a nd conflicts resolved
