+++ lib/common-lib/src/services/benificiaryRegistoryService.js
@@ -105,3 +105,35 @@ export const statusUpdate = async (data = {}, header = {}) => {
     }
   }
 }
+
+export const listOfEnum = async (params = {}, header = {}) => {
+  let headers = {
+    Authorization: 'Bearer ' + localStorage.getItem('token'),
+    ...header
+  }
+  const result = await get(`${process.env.REACT_APP_API_URL}/enum/list`, {
+    params,
+    headers
+  })
+  if (result.data) {
+    return result.data.data
+  } else {
+    return []
+  }
+}
+
+export const lastYear = async (filters = {}, header = {}) => {
+  let headers = {
+    Authorization: 'Bearer ' + localStorage.getItem('token'),
+    ...header
+  }
+
+  const currentYear = new Date().getFullYear()
+  const startAge = 14
+  const endAge = 30
+  const years = Array.from({ length: endAge - startAge + 1 }, (_, index) => ({
+    value: currentYear - startAge - index
+  }))
+
+  return years
+}
